+++ src/configstacker/sources/inifile.py
@@ -48,7 +48,7 @@ class INIFile(base.Source):
     class Meta:
         is_typed = False
 
-    def __init__(self, source, subsection_token=None, root_section='__root__',
+    def __init__(self, source, subsection_token='.', root_section='__root__',
                  **kwargs):
         super(INIFile, self).__init__(**kwargs)
         self._source = source
@@ -61,7 +61,7 @@ class INIFile(base.Source):
         for section in self._parser.sections():
             if section == self.root_section:
                 subsections = []
-            elif self.subsection_token and self.subsection_token in section:
+            elif self.subsection_token in section:
                 subsections = section.split(self.subsection_token)
             else:
                 subsections = [section]