+++ tests/sources/test_inifile.py
@@ -30,7 +30,7 @@ def test_read_ini_source(ini_file):
 
     assert config.a == '1'
     assert config.b.c == '2'
-    assert config['b.d'].e == '3'
+    assert config['b']['d'].e == '3'
     assert config['b/d/f'].g == '4'
 
 
@@ -40,7 +40,7 @@ def test_read_ini_source_from_file_object(ini_file):
 
     assert config.a == '1'
     assert config.b.c == '2'
-    assert config['b.d'].e == '3'
+    assert config['b']['d'].e == '3'
     assert config['b/d/f'].g == '4'
 
 
@@ -49,7 +49,7 @@ def test_read_ini_source_with_subsections(ini_file):
 
     assert config.a == '1'
     assert config.b.c == '2'
-    assert config.b.d.e == '3'
+    assert config['b']['d'].e == '3'
     assert config['b/d/f'].g == '4'
 
 
@@ -58,12 +58,12 @@ def test_write_ini_source(ini_file):
 
     assert config.a == '1'
     assert config.b.c == '2'
-    assert config.b.d.e == '3'
+    assert config['b']['d'].e == '3'
     assert config['b/d/f'].g == '4'
 
     config.a = 10
     config.b.c = 20
-    config.b.d.e = 30
+    config['b']['d'].e = 30
     config['b/d/f'].g = 40
     config.x.y.z = 50
 
@@ -88,5 +88,5 @@ def test_change_root_name(inimaker):
 
     assert config.a == '1'
     assert config.b.c == '2'
-    assert config['b.d'].e == '3'
+    assert config['b']['d'].e == '3'
     assert config['b/d/f'].g == '4'